+++ src/main/java/com/example/personal_blog/entity/RoleUser.java
@@ -17,5 +17,5 @@ public class RoleUser {
     @Id
     private long userID;
     @Id
-    private byte roleID;
+    private int roleID;
 }
