+++ src/main/java/com/example/personal_blog/repository/RoleRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepo extends JpaRepository<Role, Byte> {
+public interface RoleRepo extends JpaRepository<Role, Integer> {
 }
